refactor: shift FIELDS_TO_REJECT to field types
