--- src/app/utils/field-validation/config.js
@@ -1,7 +1,0 @@
-const FIELDS_TO_REJECT = require('../../../shared/resources/basic')
-  .types.filter((f) => !f.submitted)
-  .map((f) => f.name)
-
-module.exports = {
-  FIELDS_TO_REJECT,
-}
